+++ packages/cli/src/commands/recipe-site-seeds.ts
@@ -1,0 +1,498 @@
+import { readFile } from "node:fs/promises"
+import { basename, resolve } from "node:path"
+import { assertFixtureImportDeterministicIdsSupported, fixtureImportDeterministicIdPlan, type ExecutionResult, type Runtime, type WorkspaceRecipe, type WorkspaceRecipeSiteSeed } from "@automattic/wp-codebox-core"
+import { stripUndefined } from "@automattic/wp-codebox-core/internals"
+import { siteSeedScopesAreBounded } from "../recipe-dry-run.js"
+import type { RecipeExecutionResult, RecipeRunSiteSeed } from "./recipe-run-types.js"
+
+export async function importRecipeSiteSeeds(recipe: WorkspaceRecipe, recipeDirectory: string, runtime: Runtime, executions: RecipeExecutionResult[]): Promise<RecipeRunSiteSeed[]> {
+  const results: RecipeRunSiteSeed[] = []
+
+  for (const [index, siteSeed] of (recipe.inputs?.siteSeeds ?? []).entries()) {
+    const base = recipeSiteSeedRunBase(siteSeed, recipeDirectory, index)
+    if (siteSeed.type !== "fixture") {
+      results.push({
+        ...base,
+        action: "skipped",
+        reason: "parent-site export is not implemented in this first executable site seed slice",
+      })
+      continue
+    }
+
+    const format = siteSeed.format ?? "json"
+    const source = resolve(recipeDirectory, siteSeed.source ?? "")
+    if (format === "json") {
+      const rawSeed = JSON.parse(await readFile(source, "utf8"))
+      assertFixtureImportDeterministicIdsSupported(siteSeed, rawSeed)
+      const bounded = boundedFixtureSeed(rawSeed, siteSeed.scopes)
+      const execution = await runtime.execute({
+        command: "wordpress.run-php",
+        args: [`code=${siteSeedJsonImportCode(siteSeed.name, bounded.seed)}`],
+      })
+      executions.push(withRecipeExecutionPhase(execution, "setup", index))
+      const imported = parseSiteSeedImportResult(execution.stdout)
+      results.push({
+        ...base,
+        action: "imported",
+        counts: {
+          ...bounded.counts,
+          ...imported.counts,
+        },
+        ...(imported.warnings.length > 0 ? { warnings: imported.warnings } : {}),
+        ...(siteSeed.deterministicIds ? { deterministicIds: fixtureImportDeterministicIdPlan(siteSeed, rawSeed) } : {}),
+        provenance: {
+          importer: "json",
+          source,
+          ...(imported.provenance ?? {}),
+        },
+      })
+      continue
+    }
+
+    const sourceContents = await readFile(source, "utf8")
+    assertFixtureImportDeterministicIdsSupported(siteSeed)
+    const execution = await runtime.execute({
+      command: "wordpress.run-php",
+      args: [`code=${siteSeedRegistryImportCode(siteSeed, format, source, sourceContents)}`],
+    })
+    executions.push(withRecipeExecutionPhase(execution, "setup", index))
+    const imported = parseSiteSeedImportResult(execution.stdout)
+    results.push({
+      ...base,
+      action: "imported",
+      counts: imported.counts,
+      ...(imported.warnings.length > 0 ? { warnings: imported.warnings } : {}),
+      provenance: {
+        importer: format,
+        source,
+        ...(imported.provenance ?? {}),
+      },
+    })
+  }
+
+  return results
+}
+
+function withRecipeExecutionPhase(execution: ExecutionResult, recipePhase: "setup", recipeStepIndex: number): RecipeExecutionResult {
+  return {
+    ...execution,
+    recipePhase,
+    recipeStepIndex,
+  }
+}
+
+function recipeSiteSeedRunBase(siteSeed: WorkspaceRecipeSiteSeed, recipeDirectory: string, index: number): Omit<RecipeRunSiteSeed, "action" | "reason" | "counts"> {
+  return {
+    index,
+    type: siteSeed.type,
+    name: siteSeed.name,
+    ...(siteSeed.source ? { source: resolve(recipeDirectory, siteSeed.source) } : {}),
+    ...(siteSeed.format ? { format: siteSeed.format } : {}),
+    ...(siteSeed.type === "fixture" ? { importer: siteSeed.format ?? "json" } : {}),
+    ...(siteSeed.deterministicIds ? { deterministicIds: fixtureImportDeterministicIdPlan(siteSeed) } : {}),
+    ...(siteSeed.bootstrap ? { bootstrap: siteSeed.bootstrap } : {}),
+    scopes: siteSeed.scopes,
+    bounded: siteSeedScopesAreBounded(siteSeed),
+    privacy: {
+      exportsParentSiteData: false,
+      importsIntoSandbox: siteSeed.type === "fixture",
+      includesRecordData: siteSeed.type === "fixture",
+      secrets: "excluded-by-default",
+    },
+  }
+}
+
+function parseSiteSeedImportResult(stdout: string): { counts: Record<string, number>; warnings: string[]; provenance?: Record<string, unknown> } {
+  const parsed = JSON.parse(stdout.trim() || "{}") as { counts?: Record<string, unknown>; warnings?: unknown[]; provenance?: unknown }
+  const counts: Record<string, number> = {}
+  for (const [key, value] of Object.entries(parsed.counts ?? {})) {
+    if (typeof value === "number") {
+      counts[key] = value
+    }
+  }
+  const warnings = (parsed.warnings ?? []).filter((warning): warning is string => typeof warning === "string")
+  return {
+    counts,
+    warnings,
+    ...(parsed.provenance && typeof parsed.provenance === "object" && !Array.isArray(parsed.provenance) ? { provenance: parsed.provenance as Record<string, unknown> } : {}),
+  }
+}
+
+function siteSeedJsonImportCode(seedName: string, seed: unknown): string {
+  const encodedSeed = JSON.stringify(JSON.stringify(seed))
+  const encodedName = JSON.stringify(seedName)
+  return `
+$seed_name = ${encodedName};
+$seed = json_decode(${encodedSeed}, true);
+if (!is_array($seed)) {
+    throw new RuntimeException('Site seed fixture must decode to a JSON object.');
+}
+
+$counts = array('posts' => 0, 'options' => 0, 'terms' => 0, 'users' => 0, 'media' => 0, 'activePlugins' => 0, 'activeTheme' => 0);
+
+foreach (($seed['posts'] ?? array()) as $post) {
+    if (!is_array($post)) {
+        continue;
+    }
+    $postarr = array(
+        'post_type' => isset($post['post_type']) ? (string) $post['post_type'] : 'post',
+        'post_status' => isset($post['post_status']) ? (string) $post['post_status'] : (isset($post['status']) ? (string) $post['status'] : 'publish'),
+        'post_title' => isset($post['post_title']) ? (string) $post['post_title'] : (isset($post['title']) ? (string) $post['title'] : 'Seeded post'),
+        'post_content' => isset($post['post_content']) ? (string) $post['post_content'] : (isset($post['content']) ? (string) $post['content'] : ''),
+        'post_excerpt' => isset($post['post_excerpt']) ? (string) $post['post_excerpt'] : (isset($post['excerpt']) ? (string) $post['excerpt'] : ''),
+    );
+    if (isset($post['slug'])) {
+        $postarr['post_name'] = (string) $post['slug'];
+    } elseif (isset($post['post_name'])) {
+        $postarr['post_name'] = (string) $post['post_name'];
+    }
+    $post_id = wp_insert_post($postarr, true);
+    if (is_wp_error($post_id)) {
+        throw new RuntimeException('Failed to import site seed post from ' . $seed_name . ': ' . $post_id->get_error_message());
+    }
+    $counts['posts']++;
+}
+
+$options = $seed['options'] ?? array();
+if (is_array($options)) {
+    foreach ($options as $key => $option) {
+        if (is_array($option) && array_key_exists('name', $option)) {
+            update_option((string) $option['name'], $option['value'] ?? '');
+            $counts['options']++;
+            continue;
+        }
+        if (is_string($key)) {
+            update_option($key, $option);
+            $counts['options']++;
+        }
+    }
+}
+
+foreach (($seed['terms'] ?? array()) as $term) {
+    if (!is_array($term) || empty($term['name']) || empty($term['taxonomy'])) {
+        continue;
+    }
+    $result = wp_insert_term((string) $term['name'], (string) $term['taxonomy'], array_filter(array(
+        'slug' => isset($term['slug']) ? (string) $term['slug'] : null,
+        'description' => isset($term['description']) ? (string) $term['description'] : null,
+    ), static fn($value) => $value !== null));
+    if (is_wp_error($result) && 'term_exists' !== $result->get_error_code()) {
+        throw new RuntimeException('Failed to import site seed term from ' . $seed_name . ': ' . $result->get_error_message());
+    }
+    $counts['terms']++;
+}
+
+foreach (($seed['users'] ?? array()) as $user) {
+    if (!is_array($user)) {
+        continue;
+    }
+    $login = isset($user['user_login']) ? (string) $user['user_login'] : (isset($user['login']) ? (string) $user['login'] : '');
+    if ('' === $login || !preg_match('/^[A-Za-z0-9_.@-]+$/', $login)) {
+        throw new RuntimeException('Unsafe site seed user login from ' . $seed_name . '.');
+    }
+    if (username_exists($login)) {
+        $counts['users']++;
+        continue;
+    }
+    $email = isset($user['user_email']) ? (string) $user['user_email'] : (isset($user['email']) ? (string) $user['email'] : $login . '@example.invalid');
+    if (!is_email($email)) {
+        $email = $login . '@example.invalid';
+    }
+    $user_id = wp_insert_user(array(
+        'user_login' => $login,
+        'user_pass' => wp_generate_password(24, true, true),
+        'user_email' => $email,
+        'display_name' => isset($user['display_name']) ? (string) $user['display_name'] : $login,
+        'role' => isset($user['role']) ? (string) $user['role'] : (is_array($user['roles'] ?? null) && count($user['roles']) > 0 ? (string) reset($user['roles']) : 'subscriber'),
+    ));
+    if (is_wp_error($user_id)) {
+        throw new RuntimeException('Failed to import site seed user from ' . $seed_name . ': ' . $user_id->get_error_message());
+    }
+    $counts['users']++;
+}
+
+foreach (($seed['media'] ?? array()) as $media) {
+    if (!is_array($media)) {
+        continue;
+    }
+    $attachment = array(
+        'post_type' => 'attachment',
+        'post_status' => isset($media['post_status']) ? (string) $media['post_status'] : 'inherit',
+        'post_title' => isset($media['post_title']) ? (string) $media['post_title'] : (isset($media['title']) ? (string) $media['title'] : 'Seeded media'),
+        'post_content' => isset($media['post_content']) ? (string) $media['post_content'] : '',
+        'post_excerpt' => isset($media['post_excerpt']) ? (string) $media['post_excerpt'] : '',
+        'post_mime_type' => isset($media['post_mime_type']) ? (string) $media['post_mime_type'] : (isset($media['mime_type']) ? (string) $media['mime_type'] : ''),
+    );
+    if (isset($media['slug'])) {
+        $attachment['post_name'] = (string) $media['slug'];
+    } elseif (isset($media['post_name'])) {
+        $attachment['post_name'] = (string) $media['post_name'];
+    }
+    $attachment_id = wp_insert_post($attachment, true);
+    if (is_wp_error($attachment_id)) {
+        throw new RuntimeException('Failed to import site seed media from ' . $seed_name . ': ' . $attachment_id->get_error_message());
+    }
+    $counts['media']++;
+}
+
+$active_plugins = $seed['activePlugins'] ?? array();
+if (is_array($active_plugins) && count($active_plugins) > 0) {
+    require_once ABSPATH . 'wp-admin/includes/plugin.php';
+    foreach ($active_plugins as $plugin) {
+        $plugin_file = is_array($plugin) ? ($plugin['pluginFile'] ?? ($plugin['file'] ?? '')) : $plugin;
+        $plugin_file = is_string($plugin_file) ? $plugin_file : '';
+        if ('' === $plugin_file || str_starts_with($plugin_file, '/') || str_contains($plugin_file, '..') || !str_ends_with($plugin_file, '.php')) {
+            throw new RuntimeException('Unsafe site seed active plugin entry from ' . $seed_name . '.');
+        }
+        if (!file_exists(WP_PLUGIN_DIR . '/' . $plugin_file)) {
+            throw new RuntimeException('Site seed active plugin is not installed in sandbox: ' . $plugin_file);
+        }
+        $result = activate_plugin($plugin_file, '', false, true);
+        if (is_wp_error($result)) {
+            throw new RuntimeException('Failed to activate site seed plugin from ' . $seed_name . ': ' . $result->get_error_message());
+        }
+        $counts['activePlugins']++;
+    }
+}
+
+$active_theme = $seed['activeTheme'] ?? null;
+if (is_array($active_theme)) {
+    $active_theme = $active_theme['stylesheet'] ?? ($active_theme['slug'] ?? null);
+}
+if (is_string($active_theme) && '' !== $active_theme) {
+    if (!preg_match('/^[A-Za-z0-9_-]+$/', $active_theme)) {
+        throw new RuntimeException('Unsafe site seed active theme entry from ' . $seed_name . '.');
+    }
+    $theme = wp_get_theme($active_theme);
+    if (!$theme->exists()) {
+        throw new RuntimeException('Site seed active theme is not installed in sandbox: ' . $active_theme);
+    }
+    switch_theme($active_theme);
+    $counts['activeTheme']++;
+}
+
+echo wp_json_encode(array('schema' => 'wp-codebox/site-seed-import/v1', 'name' => $seed_name, 'counts' => $counts));
+`}
+
+function siteSeedRegistryImportCode(siteSeed: WorkspaceRecipeSiteSeed, format: string, source: string, sourceContents: string): string {
+  const encodedName = JSON.stringify(siteSeed.name)
+  const encodedFormat = JSON.stringify(format)
+  const encodedSource = JSON.stringify(source)
+  const encodedSourceBasename = JSON.stringify(basename(source))
+  const encodedSourceContents = JSON.stringify(sourceContents)
+  const encodedScopes = JSON.stringify(JSON.stringify(siteSeed.scopes))
+  return `
+$seed_name = ${encodedName};
+$format = ${encodedFormat};
+$source = ${encodedSource};
+$source_basename = ${encodedSourceBasename};
+$source_contents = ${encodedSourceContents};
+$scopes = json_decode(${encodedScopes}, true);
+if (!is_array($scopes)) {
+    throw new RuntimeException('Site seed scopes must decode to an object.');
+}
+
+$importers = apply_filters('wp_codebox_site_seed_importers', array());
+if (!is_array($importers) || !array_key_exists($format, $importers)) {
+    throw new RuntimeException('No WP Codebox site seed importer registered for format: ' . $format);
+}
+
+$importer = $importers[$format];
+$callback = is_array($importer) && array_key_exists('callback', $importer) ? $importer['callback'] : $importer;
+if (!is_callable($callback)) {
+    throw new RuntimeException('WP Codebox site seed importer is not callable for format: ' . $format);
+}
+
+$result = call_user_func($callback, array(
+    'schema' => 'wp-codebox/site-seed-import-request/v1',
+    'name' => $seed_name,
+    'format' => $format,
+    'source' => $source,
+    'source_basename' => $source_basename,
+    'source_contents' => $source_contents,
+    'scopes' => $scopes,
+    'metadata' => array(
+        'source_size' => strlen($source_contents),
+    ),
+));
+
+if (!is_array($result)) {
+    throw new RuntimeException('WP Codebox site seed importer must return an array for format: ' . $format);
+}
+
+$counts = array();
+foreach (($result['counts'] ?? array()) as $key => $value) {
+    if (is_int($value) || is_float($value)) {
+        $counts[(string) $key] = $value;
+    }
+}
+
+$warnings = array();
+foreach (($result['warnings'] ?? array()) as $warning) {
+    if (is_string($warning)) {
+        $warnings[] = $warning;
+    }
+}
+
+$provenance = isset($result['provenance']) && is_array($result['provenance']) ? $result['provenance'] : array();
+$provenance['importer'] = $format;
+$provenance['source'] = $source;
+
+echo wp_json_encode(array(
+    'schema' => 'wp-codebox/site-seed-import/v1',
+    'name' => $seed_name,
+    'importer' => $format,
+    'counts' => $counts,
+    'warnings' => $warnings,
+    'provenance' => $provenance,
+));
+`}
+
+function boundedFixtureSeed(rawSeed: unknown, scopes: WorkspaceRecipeSiteSeed["scopes"]): { seed: Record<string, unknown>; counts: Record<string, number> } {
+  if (!rawSeed || typeof rawSeed !== "object" || Array.isArray(rawSeed)) {
+    throw new Error("Recipe fixture siteSeed JSON must be an object")
+  }
+
+  const seed = rawSeed as Record<string, unknown>
+  const posts = boundedRecords(arrayRecords(seed.posts), scopes.posts, (record, scope) => matchesPostScope(record, scope))
+  const options = boundedOptions(seed.options, scopes.options)
+  const terms = boundedRecords(arrayRecords(seed.terms), scopes.terms, (record, scope) => matchesTermScope(record, scope))
+  const users = boundedRecords(arrayRecords(seed.users), scopes.users, (record, scope) => matchesUserScope(record, scope))
+  const media = boundedRecords(arrayRecords(seed.media), scopes.media, (record, scope) => matchesMediaScope(record, scope))
+  const activePlugins = boundedActivePlugins(seed.activePlugins, scopes.activePlugins)
+  const activeTheme = boundedActiveTheme(seed.activeTheme, scopes.activeTheme)
+
+  return {
+    seed: stripUndefined({ posts: posts.records, options: options.records, terms: terms.records, users: users.records, media: media.records, activePlugins: activePlugins.records, activeTheme: activeTheme.record }),
+    counts: {
+      fixturePostsIncluded: posts.records.length,
+      fixturePostsExcluded: posts.excluded,
+      fixtureOptionsIncluded: options.count,
+      fixtureOptionsExcluded: options.excluded,
+      fixtureTermsIncluded: terms.records.length,
+      fixtureTermsExcluded: terms.excluded,
+      fixtureUsersIncluded: users.records.length,
+      fixtureUsersExcluded: users.excluded,
+      fixtureMediaIncluded: media.records.length,
+      fixtureMediaExcluded: media.excluded,
+      fixtureActivePluginsIncluded: activePlugins.records.length,
+      fixtureActivePluginsExcluded: activePlugins.excluded,
+      fixtureActiveThemeIncluded: activeTheme.record === undefined ? 0 : 1,
+      fixtureActiveThemeExcluded: activeTheme.excluded,
+    },
+  }
+}
+
+function boundedActivePlugins(activePlugins: unknown, scope: boolean | undefined): { records: Array<string | Record<string, unknown>>; excluded: number } {
+  const records = Array.isArray(activePlugins)
+    ? activePlugins.filter((plugin): plugin is string | Record<string, unknown> => typeof plugin === "string" || (Boolean(plugin) && typeof plugin === "object" && !Array.isArray(plugin)))
+    : []
+
+  if (scope !== true) {
+    return { records: [], excluded: records.length }
+  }
+
+  const included = records.slice(0, 100)
+  return { records: included, excluded: records.length - included.length }
+}
+
+function boundedActiveTheme(activeTheme: unknown, scope: boolean | undefined): { record: unknown; excluded: number } {
+  if (activeTheme === undefined || activeTheme === null) {
+    return { record: undefined, excluded: 0 }
+  }
+
+  if (scope !== true) {
+    return { record: undefined, excluded: 1 }
+  }
+
+  return { record: activeTheme, excluded: 0 }
+}
+
+function arrayRecords(value: unknown): Array<Record<string, unknown>> {
+  return Array.isArray(value) ? value.filter((item): item is Record<string, unknown> => Boolean(item) && typeof item === "object" && !Array.isArray(item)) : []
+}
+
+function boundedRecords(records: Array<Record<string, unknown>>, scope: WorkspaceRecipeSiteSeed["scopes"]["posts"], matches: (record: Record<string, unknown>, scope: NonNullable<WorkspaceRecipeSiteSeed["scopes"]["posts"]>) => boolean): { records: Array<Record<string, unknown>>; excluded: number } {
+  if (!scope) {
+    return { records: [], excluded: records.length }
+  }
+
+  const filtered = records.filter((record) => matches(record, scope))
+  const maxRecords = scope.maxRecords ?? filtered.length
+  return {
+    records: filtered.slice(0, maxRecords),
+    excluded: records.length - Math.min(filtered.length, maxRecords),
+  }
+}
+
+function boundedOptions(options: unknown, scope: WorkspaceRecipeSiteSeed["scopes"]["options"]): { records: Record<string, unknown> | Array<Record<string, unknown>> | undefined; count: number; excluded: number } {
+  if (!scope || !scope.names || scope.names.length === 0) {
+    const count = Array.isArray(options) ? options.length : options && typeof options === "object" ? Object.keys(options).length : 0
+    return { records: undefined, count: 0, excluded: count }
+  }
+
+  const allowed = new Set(scope.names)
+  const maxRecords = scope.maxRecords ?? allowed.size
+  if (Array.isArray(options)) {
+    const filtered = options.filter((option): option is Record<string, unknown> => Boolean(option) && typeof option === "object" && !Array.isArray(option) && typeof option.name === "string" && allowed.has(option.name)).slice(0, maxRecords)
+    return { records: filtered, count: filtered.length, excluded: options.length - filtered.length }
+  }
+
+  if (!options || typeof options !== "object") {
+    return { records: undefined, count: 0, excluded: 0 }
+  }
+
+  const entries = Object.entries(options as Record<string, unknown>).filter(([name]) => allowed.has(name)).slice(0, maxRecords)
+  return { records: Object.fromEntries(entries), count: entries.length, excluded: Object.keys(options).length - entries.length }
+}
+
+function matchesPostScope(record: Record<string, unknown>, scope: NonNullable<WorkspaceRecipeSiteSeed["scopes"]["posts"]>): boolean {
+  return matchesNumberSelector(record, scope.ids, ["id", "ID"]) &&
+    matchesStringSelector(record, scope.slugs, ["slug", "post_name"]) &&
+    matchesStringSelector(record, scope.postTypes, ["post_type", "type"]) &&
+    matchesStringSelector(record, scope.statuses, ["post_status", "status"])
+}
+
+function matchesTermScope(record: Record<string, unknown>, scope: NonNullable<WorkspaceRecipeSiteSeed["scopes"]["terms"]>): boolean {
+  return matchesNumberSelector(record, scope.ids, ["id", "term_id"]) &&
+    matchesStringSelector(record, scope.slugs, ["slug"]) &&
+    matchesStringSelector(record, scope.names, ["name"]) &&
+    matchesStringSelector(record, scope.taxonomies, ["taxonomy"])
+}
+
+function matchesUserScope(record: Record<string, unknown>, scope: NonNullable<WorkspaceRecipeSiteSeed["scopes"]["users"]>): boolean {
+  return matchesNumberSelector(record, scope.ids, ["id", "ID"]) &&
+    matchesStringSelector(record, scope.names, ["user_login", "login", "display_name", "name"]) &&
+    matchesArrayStringSelector(record, scope.roles, ["roles"])
+}
+
+function matchesMediaScope(record: Record<string, unknown>, scope: NonNullable<WorkspaceRecipeSiteSeed["scopes"]["media"]>): boolean {
+  return matchesNumberSelector(record, scope.ids, ["id", "ID"]) &&
+    matchesStringSelector(record, scope.slugs, ["slug", "post_name"]) &&
+    matchesStringSelector(record, scope.names, ["post_title", "title", "name"]) &&
+    matchesStringSelector(record, scope.statuses, ["post_status", "status"])
+}
+
+function matchesStringSelector(record: Record<string, unknown>, allowed: string[] | undefined, keys: string[]): boolean {
+  if (!allowed || allowed.length === 0) {
+    return true
+  }
+  const values = keys.map((key) => record[key]).filter((value): value is string => typeof value === "string")
+  return values.some((value) => allowed.includes(value))
+}
+
+function matchesArrayStringSelector(record: Record<string, unknown>, allowed: string[] | undefined, keys: string[]): boolean {
+  if (!allowed || allowed.length === 0) {
+    return true
+  }
+  const values = keys.flatMap((key) => Array.isArray(record[key]) ? record[key] : []).filter((value): value is string => typeof value === "string")
+  return values.some((value) => allowed.includes(value))
+}
+
+function matchesNumberSelector(record: Record<string, unknown>, allowed: number[] | undefined, keys: string[]): boolean {
+  if (!allowed || allowed.length === 0) {
+    return true
+  }
+  const values = keys.map((key) => record[key]).filter((value): value is number => typeof value === "number")
+  return values.some((value) => allowed.includes(value))
+}
